+++ influxql/parser_test.go
@@ -718,12 +718,8 @@ func TestParser_ParseStatement(t *testing.T) {
 
 		// DROP SERIES statement
 		{
-			s:    `DROP SERIES 1`,
-			stmt: &influxql.DropSeriesStatement{SeriesID: 1},
-		},
-		{
 			s:    `DROP SERIES FROM src`,
-			stmt: &influxql.DropSeriesStatement{Source: &influxql.Measurement{Name: "src"}},
+			stmt: &influxql.DropSeriesStatement{Sources: []influxql.Source{&influxql.Measurement{Name: "src"}}},
 		},
 		{
 			s: `DROP SERIES WHERE host = 'hosta.influxdb.org'`,
@@ -738,7 +734,7 @@ func TestParser_ParseStatement(t *testing.T) {
 		{
 			s: `DROP SERIES FROM src WHERE host = 'hosta.influxdb.org'`,
 			stmt: &influxql.DropSeriesStatement{
-				Source: &influxql.Measurement{Name: "src"},
+				Sources: []influxql.Source{&influxql.Measurement{Name: "src"}},
 				Condition: &influxql.BinaryExpr{
 					Op:  influxql.EQ,
 					LHS: &influxql.VarRef{Val: "host"},
@@ -1158,7 +1154,7 @@ func TestParser_ParseStatement(t *testing.T) {
 		{s: `DELETE FROM`, err: `found EOF, expected identifier at line 1, char 13`},
 		{s: `DELETE FROM myseries WHERE`, err: `found EOF, expected identifier, string, number, bool at line 1, char 28`},
 		{s: `DROP MEASUREMENT`, err: `found EOF, expected identifier at line 1, char 18`},
-		{s: `DROP SERIES`, err: `found EOF, expected number at line 1, char 13`},
+		{s: `DROP SERIES`, err: `found EOF, expected FROM, WHERE at line 1, char 13`},
 		{s: `DROP SERIES FROM`, err: `found EOF, expected identifier at line 1, char 18`},
 		{s: `DROP SERIES FROM src WHERE`, err: `found EOF, expected identifier, string, number, bool at line 1, char 28`},
 		{s: `SHOW CONTINUOUS`, err: `found EOF, expected QUERIES at line 1, char 17`},
@@ -1531,17 +1527,13 @@ func TestDropSeriesStatement_String(t *testing.T) {
 		stmt influxql.Statement
 	}{
 		{
-			s:    `DROP SERIES 1`,
-			stmt: &influxql.DropSeriesStatement{SeriesID: 1},
-		},
-		{
 			s:    `DROP SERIES FROM src`,
-			stmt: &influxql.DropSeriesStatement{Source: &influxql.Measurement{Name: "src"}},
+			stmt: &influxql.DropSeriesStatement{Sources: []influxql.Source{&influxql.Measurement{Name: "src"}}},
 		},
 		{
 			s: `DROP SERIES FROM src WHERE host = 'hosta.influxdb.org'`,
 			stmt: &influxql.DropSeriesStatement{
-				Source: &influxql.Measurement{Name: "src"},
+				Sources: []influxql.Source{&influxql.Measurement{Name: "src"}},
 				Condition: &influxql.BinaryExpr{
 					Op:  influxql.EQ,
 					LHS: &influxql.VarRef{Val: "host"},
@@ -1552,7 +1544,7 @@ func TestDropSeriesStatement_String(t *testing.T) {
 		{
 			s: `DROP SERIES FROM src WHERE host = 'hosta.influxdb.org'`,
 			stmt: &influxql.DropSeriesStatement{
-				Source: &influxql.Measurement{Name: "src"},
+				Sources: []influxql.Source{&influxql.Measurement{Name: "src"}},
 				Condition: &influxql.BinaryExpr{
 					Op:  influxql.EQ,
 					LHS: &influxql.VarRef{Val: "host"},